requirements.txt
rich>=13.7.0
python-dotenv>=1.0.0

